src/requirements/base.txt
# Core
Django>=1.7,<1.8
psycopg2>=2.6,<2.7

# Packages
django-extensions
djangorestframework-jwt==1.5.0
djangorestframework>=3.0,<3.1
ipython
Werkzeug
